Add files via upload
This repository hosts the supplementary materials for the paper titled "Coordinated Dispatch of Hydrogen-Penetrated Integrated Gas-Electricity System Based on Dynamic Subsidy: From a Multi-Stakeholder View". These supplementary materials include the complete detailed derivation processes that support the analyses and results presented in the paper.

Contents:

Detailed Derivations: Comprehensive documents that provide step-by-step derivations of the formulas and theories discussed in the paper. These are essential for understanding the mathematical foundations and the analytical approach used in our research.
These materials are made available to provide transparency and enhance the reproducibility of our work. Researchers and readers are encouraged to delve into these detailed derivations to better understand the complexities and rigor of our study.

How to Cite This Material: If you utilize or refer to these materials in your research, please cite them as follows:

[Authors]. "Supplementary Materials for 'XXXX'". GitHub, [year]. Available at [URL to repository]